pyproject.toml
[tool.poetry]
name = "fastapi-html"
version = "0.1.0"
description = "FastAPI with HTMX"
authors = ["Michael Bailey <mbailey1302@gmail.com>"]
license = "Propietary"

[tool.poetry.dependencies]
python = "^3.10"
structlog = "^23.1.0"
python-jose = "^3.3.0"
uvicorn = "^0.23.1"
gunicorn = "^21.2.0"
beanie = "^1.20.0"
asgi-correlation-id = "^4.2.0"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
pydantic = {version = "^2.0.3", extras = ["email"]}
pydantic-settings = "^2.0.2"
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"
fastapi = "^0.103.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
isort = "^5.12.0"
flake8 = "^6.0.0"
faker = "^17.6.0"
black = "^23.3.0"
mypy = "^1.2.0"
httpx = "^0.24.1"

[tool.black]
line-length=79

[tool.isort]
profile="black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
